# docs/open-api/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "DiffScope Synthesis Platform API",
    "version": "0.0.0",
    "description": "DSSP HTTP API surface represented by this service. A server implementation can support a subset of the broader DSSP API."
  },
  "servers": [
    {
      "url": "http://localhost:13711",
      "description": "Default local server URL. The actual host and port are configured by the service."
    }
  ],
...
  "paths": {
...
    "/arch/{arch_id}": {
      "get": {
        "tags": [
          "Architecture and Singer"
        ],
        "operationId": "getArchitecture",
        "summary": "Get architecture metadata",
        "description": "Returns metadata for one registered architecture.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ArchID"
          },
          {
            "$ref": "#/components/parameters/DisplayLanguage"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Architecture"
          }
        }
      }
    },
    "/singer": {
      "get": {
        "tags": [
          "Architecture and Singer"
        ],
        "operationId": "getSingerList",
        "summary": "List all singers",
        "description": "Returns singers from every registered architecture.",
        "parameters": [
          {
            "$ref": "#/components/parameters/DisplayLanguage"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SingerList"
          }
        }
      }
    },
    "/arch/{arch_id}/singer": {
      "get": {
        "tags": [
          "Architecture and Singer"
        ],
        "operationId": "getArchSingerList",
        "summary": "List singers for an architecture",
        "description": "Returns singers from one registered architecture.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ArchID"
          },
          {
            "$ref": "#/components/parameters/DisplayLanguage"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ArchSingerList"
          }
        }
      }
    },
    "/arch/{arch_id}/singer/{singer_id}": {
      "get": {
        "tags": [
          "Architecture and Singer"
        ],
        "operationId": "getArchSinger",
        "summary": "Get singer metadata",
        "description": "Returns one singer from one registered architecture.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ArchID"
          },
          {
            "$ref": "#/components/parameters/SingerID"
          },
          {
            "$ref": "#/components/parameters/DisplayLanguage"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Singer"
          }
        }
      }
    },
    "/arch/{arch_id}/singer/{singer_id}/avatar": {
      "get": {
        "tags": [
          "Architecture and Singer"
        ],
        "operationId": "getArchSingerAvatar",
        "summary": "Get singer avatar URL",
        "description": "Returns the avatar URL for one singer.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ArchID"
          },
          {
            "$ref": "#/components/parameters/SingerID"
          },
          {
            "$ref": "#/components/parameters/DisplayLanguage"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SingerAvatar"
          }
        }
      }
    },
    "/arch/{arch_id}/singer/{singer_id}/background": {
      "get": {
        "tags": [
          "Architecture and Singer"
        ],
        "operationId": "getArchSingerBackground",
        "summary": "Get singer background URL",
        "description": "Returns the background image URL for one singer.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ArchID"
          },
          {
            "$ref": "#/components/parameters/SingerID"
          },
          {
            "$ref": "#/components/parameters/DisplayLanguage"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SingerBackground"
          }
        }
      }
    },
    "/arch/{arch_id}/singer/{singer_id}/demo_audio": {
      "get": {
        "tags": [
          "Architecture and Singer"
        ],
        "operationId": "getArchSingerDemoAudioList",
        "summary": "List singer demo audio",
        "description": "Returns demo audio entries for one singer.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ArchID"
          },
          {
            "$ref": "#/components/parameters/SingerID"
          },
          {
            "$ref": "#/components/parameters/DisplayLanguage"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SingerDemoAudioList"
          }
        }
      }
    },
...
    "/synth/parameter": {
      "post": {
        "tags": [
          "Synthesis"
        ],
        "operationId": "postSynthParameter",
        "summary": "Generate parameter values",
        "description": "Generates requested parameter values. Set stream to true to receive newline-delimited JSON events.",
        "requestBody": {
          "$ref": "#/components/requestBodies/ParameterRequest"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/ParameterResult"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "ArchID": {
        "name": "arch_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Architecture identifier."
      },
      "SingerID": {
        "name": "singer_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Singer identifier understood by the selected architecture."
      },
      "DisplayLanguage": {
        "name": "display_language",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Preferred language for display metadata."
      }
    },
    "requestBodies": {
...
      "ParameterRequest": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ParameterRequest"
            }
          }
        }
      },
...
    },
    "responses": {
...
      "Architecture": {
        "description": "Architecture metadata.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ArchitectureMetadata"
            }
          }
        }
      },
      "SingerList": {
        "description": "Singer metadata from every registered architecture.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SingerInfoList"
            }
          }
        }
      },
      "ArchSingerList": {
        "description": "Singer metadata from one registered architecture.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SingerInfoList"
            }
          }
        }
      },
      "Singer": {
        "description": "Singer metadata.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SingerInfo"
            }
          }
        }
      },
      "SingerAvatar": {
        "description": "Singer avatar URL.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SingerAvatarResponse"
            }
          }
        }
      },
      "SingerBackground": {
        "description": "Singer background URL.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SingerBackgroundResponse"
            }
          }
        }
      },
      "SingerDemoAudioList": {
        "description": "Singer demo audio entries.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SingerDemoAudioList"
            }
          }
        }
      },
...
      "ParameterResult": {
        "description": "Parameter generation result. Non-streaming requests return a complete JSON response. Streaming requests return newline-delimited JSON progress and completion events.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ParameterResponse"
            }
          },
          "application/x-ndjson": {
            "itemSchema": {
              "$ref": "#/components/schemas/ParameterStreamEvent"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ArchitectureMetadata": {
        "type": "object",
        "required": [
          "id",
          "name",
          "pronunciation_mode",
          "phoneme_mode",
          "parameters",
          "audio_dependencies"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Architecture identifier."
          },
          "name": {
            "type": "string"
          },
          "pronunciation_mode": {
            "$ref": "#/components/schemas/PronunciationMode"
          },
          "phoneme_mode": {
            "$ref": "#/components/schemas/PhonemeMode"
          },
          "parameters": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/ArchitectureParameterMetadata"
            }
          },
          "audio_dependencies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "PronunciationMode": {
        "type": "string",
        "enum": [
          "FULL",
          "SKIP"
        ]
      },
      "PhonemeMode": {
        "type": "string",
        "enum": [
          "FULL",
          "TOKEN_ONLY",
          "SKIP"
        ]
      },
      "ArchitectureParameterMetadata": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/DirectArchitectureParameterMetadata"
          },
          {
            "$ref": "#/components/schemas/IndirectArchitectureParameterMetadata"
          }
        ]
      },
      "DirectArchitectureParameterMetadata": {
        "type": "object",
        "required": [
          "type"
        ],
        "properties": {
          "type": {
            "const": "DIRECT"
          }
        }
      },
      "IndirectArchitectureParameterMetadata": {
        "type": "object",
        "required": [
          "type",
          "depends_on"
        ],
        "properties": {
          "type": {
            "const": "INDIRECT"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "SingerInfoList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/SingerInfo"
        }
      },
      "SingerInfo": {
        "type": "object",
        "required": [
          "id",
          "name",
          "arch",
          "mix_group",
          "languages",
          "default_language",
          "arch_specific_info",
          "default_extra"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Singer identifier understood by the selected architecture."
          },
          "name": {
            "type": "string"
          },
          "arch": {
            "type": "string",
            "description": "Architecture identifier."
          },
          "mix_group": {
            "type": "string",
            "description": "Identifier shared by singers whose inference resources can be mixed."
          },
          "languages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "default_language": {
            "type": "string"
          },
          "arch_specific_info": {
            "description": "Architecture-specific singer JSON payload."
          },
          "default_extra": {
            "description": "Default architecture-specific singer JSON payload."
          }
        }
      },
      "SingerAvatarResponse": {
        "type": "object",
        "required": [
          "avatar_url"
        ],
        "properties": {
          "avatar_url": {
            "type": "string"
          }
        }
      },
      "SingerBackgroundResponse": {
        "type": "object",
        "required": [
          "background_url"
        ],
        "properties": {
          "background_url": {
            "type": "string"
          }
        }
      },
      "SingerDemoAudioList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/SingerDemoAudio"
        }
      },
      "SingerDemoAudio": {
        "type": "object",
        "required": [
          "name",
          "audio_url"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "audio_url": {
            "type": "string"
          }
        }
      },
...
      "ParameterRequest": {
        "type": "object",
        "required": [
          "context",
          "input"
        ],
        "properties": {
          "context": {
            "$ref": "#/components/schemas/MultiSingerContext"
          },
          "input": {
            "$ref": "#/components/schemas/ParameterInput"
          },
          "stream": {
            "type": "boolean",
            "description": "When true, the response uses application/x-ndjson and emits progress events."
          }
        }
      },
...
      "MultiSingerContext": {
        "type": "object",
        "required": [
          "arch",
          "arch_extra",
          "singers"
        ],
        "properties": {
          "arch": {
            "type": "string",
            "description": "Architecture identifier."
          },
          "arch_extra": {
            "description": "Architecture-specific JSON payload."
          },
          "singers": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/Singer"
            }
          }
        }
      },
      "Singer": {
        "type": "object",
        "required": [
          "id",
          "extra"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Singer identifier understood by the selected architecture."
          },
          "extra": {
            "description": "Architecture-specific singer JSON payload."
          }
        }
      },
...
      "NotePosition": {
        "type": "object",
        "required": [
          "gap",
          "duration"
        ],
        "properties": {
          "gap": {
            "type": "number",
            "minimum": 0
          },
          "duration": {
            "type": "number",
            "minimum": 0
          }
        }
      },
...
      "ParameterInput": {
        "type": "object",
        "required": [
          "piece_duration",
          "notes",
          "mix",
          "mix_sample_rate",
          "parameters"
        ],
        "properties": {
          "piece_duration": {
            "type": "number",
            "minimum": 0
          },
          "notes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParameterNote"
            }
          },
          "mix": {
            "$ref": "#/components/schemas/Mix"
          },
          "mix_sample_rate": {
            "type": "number",
            "exclusiveMinimum": 0
          },
          "parameters": {
            "$ref": "#/components/schemas/ParameterMap"
          }
        }
      },
      "ParameterNote": {
        "type": "object",
        "required": [
          "position",
          "cent",
          "pronunciation",
          "language",
          "phonemes"
        ],
        "properties": {
          "position": {
            "$ref": "#/components/schemas/NotePosition"
          },
          "cent": {
            "type": "integer",
            "minimum": 0,
            "maximum": 12800
          },
          "pronunciation": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "phonemes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParameterInputPhoneme"
            }
          }
        }
      },
      "ParameterInputPhoneme": {
        "type": "object",
        "required": [
          "token",
          "onset",
          "language",
          "start"
        ],
        "properties": {
          "token": {
            "type": "string"
          },
          "onset": {
            "type": "boolean"
          },
          "language": {
            "type": "string"
          },
          "start": {
            "type": "number"
          }
        }
      },
...
      "Mix": {
        "type": "array",
        "description": "Per-frame singer mix data. Each value must be in [0, 1]. Each row must contain exactly singers.length - 1 values, and the sum of values in each row must be less than or equal to 1.",
        "items": {
          "type": "array",
          "items": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          }
        }
      },
      "ParameterMap": {
        "type": "object",
        "description": "Open string-keyed parameter map. Supported keys and semantic requirements are defined by the selected architecture.",
        "additionalProperties": {
          "$ref": "#/components/schemas/Parameter"
        }
      },
...
      "Parameter": {
        "type": "object",
        "required": [
          "values",
          "sample_rate"
        ],
        "properties": {
          "values": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "sample_rate": {
            "type": "number",
            "exclusiveMinimum": 0
          },
          "retake": {
            "$ref": "#/components/schemas/ParameterRetake"
          }
        }
      },
...
      "ParameterRetake": {
        "type": "object",
        "required": [
          "position",
          "length"
        ],
        "properties": {
          "position": {
            "type": "integer",
            "minimum": 0
          },
          "length": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "ParameterResponse": {
        "type": "object",
        "required": [
          "state",
          "output"
        ],
        "properties": {
          "state": {
            "const": "COMPLETE"
          },
          "output": {
            "$ref": "#/components/schemas/ParameterOutput"
          }
        }
      },
      "ParameterOutput": {
        "type": "object",
        "required": [
          "parameters"
        ],
        "properties": {
          "parameters": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/ParameterOutputParameter"
            }
          }
        }
      },
      "ParameterOutputParameter": {
        "type": "object",
        "required": [
          "values",
          "sample_rate"
        ],
        "properties": {
          "values": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "sample_rate": {
            "type": "number",
            "exclusiveMinimum": 0
          }
        }
      },
...
      "ParameterStreamEvent": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/StreamStateEvent"
          },
          {
            "$ref": "#/components/schemas/ParameterProcessingEvent"
          },
          {
            "$ref": "#/components/schemas/ParameterResponse"
          }
        ]
      },
...
      "StreamStateEvent": {
        "type": "object",
        "required": [
          "state"
        ],
        "properties": {
          "state": {
            "enum": [
              "QUEUING",
              "PROCESSING"
            ]
          }
        }
      },
      "ParameterProcessingEvent": {
        "type": "object",
        "required": [
          "state",
          "output"
        ],
        "properties": {
          "state": {
            "const": "PROCESSING"
          },
          "output": {
            "$ref": "#/components/schemas/ParameterOutput"
          }
        }
      }
    }
  }
}